requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
Django==3.0.5
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-post-office==3.3.1
django-six==1.0.4
importlib-metadata==1.6.0
jsonfield==3.1.0
kombu==4.6.8
pytz==2019.3
redis==3.4.1
screen==1.0.1
sqlparse==0.3.1
vine==1.3.0
xlwt==1.3.0
zipp==3.1.0
mock==4.0.2